Cleanup: Remove SHA check, use session.exec(), collapse migrations
- Remove unused SHA256 hash computation from ArcFace weight download
- Replace session.execute() with SQLModel's session.exec() across pipeline
- Collapse all migrations into single 001_initial_schema.py for alpha
  - Delete migrations 002, 003, 004
  - Update initial migration with final schema (512-dim embeddings,
    face_clusters table, updated media_person_links)
- Update e2e test to check for correct migration version
